+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add and restructure tests for `assert_equal` and `assert_equal_constant` [#733]
 - Add and restructure tests for logic component [#734]
 - Add and restructure tests for range component [#735]
-- Add and restructure integration tests for component boolean and component select [#731]
+- Add and restructure tests for boolean and select components [#731]
+- Add tests for `gate_add` and `gate_mul` [#736]
 
 ### Removed
 
@@ -468,6 +469,7 @@ is necessary since `rkyv/validation` was required as a bound.
 
 <!-- ISSUES -->
 [#746]: https://github.com/dusk-network/plonk/issues/746
+[#736]: https://github.com/dusk-network/plonk/issues/736
 [#735]: https://github.com/dusk-network/plonk/issues/735
 [#734]: https://github.com/dusk-network/plonk/issues/734
 [#743]: https://github.com/dusk-network/plonk/issues/743